Add `GameInstance` framework
Create the `GameInstance` struct, which will act as the base
struct for the `GameServer` and `GameClient` structs, respectively,
by abstracting the underlying binary into a set of functions it
is expected to export.

As a complete model for the Game Module v1 API has not yet been
finalised, it only provides the absolute minimum functionality -- I
plan to add functions to represent the additional APIs over time, on
an ad-hoc basis. The API will then be finalised/polished after an
adequate implementation has been produced.